many things

+++ ds4wizard-cpp/Bluetooth.cpp
@@ -11,7 +11,7 @@ bool Bluetooth::disconnectDevice(const gsl::span<uint8_t>& macAddress)
 
 	std::array<uint8_t, 8> buffer {};
 
-	for (int i = 0; i < macAddress.size(); i++)
+	for (size_t i = 0; i < macAddress.size(); i++)
 	{
 		buffer[5 - i] = macAddress[i];
 	}
