+++ ds4wizard-cpp/Ds4InputData.cpp
@@ -12,12 +12,12 @@ bool Ds4Stick::operator!=(const Ds4Stick& right) const
 	return !(*this == right);
 }
 
-double Ds4Stick::lengthSquared() const
+float Ds4Stick::lengthSquared() const
 {
-	return x * x + y * y;
+	return (float)x * (float)x + (float)y * (float)y;
 }
 
-double Ds4Stick::length() const
+float Ds4Stick::length() const
 {
 	return std::sqrt(lengthSquared());
 }
@@ -27,12 +27,12 @@ bool Ds4Vector2::operator==(const Ds4Vector2& other) const
 	return x == other.x && y == other.y;
 }
 
-double Ds4Vector2::lengthSquared() const
+float Ds4Vector2::lengthSquared() const
 {
-	return x * x + y * y;
+	return (float)x * (float)x + (float)y * (float)y;
 }
 
-double Ds4Vector2::length() const
+float Ds4Vector2::length() const
 {
 	return std::sqrt(lengthSquared());
 }
@@ -42,12 +42,12 @@ bool Ds4Vector3::operator==(const Ds4Vector3& other) const
 	return x == other.x && y == other.y && z == other.z;
 }
 
-double Ds4Vector3::lengthSquared() const
+float Ds4Vector3::lengthSquared() const
 {
-	return x * x + y * y + z * z;
+	return (float)x * (float)x + (float)y * (float)y + (float)z * (float)z;
 }
 
-double Ds4Vector3::length() const
+float Ds4Vector3::length() const
 {
 	return std::sqrt(lengthSquared());
 }
